% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle_side.R
\name{triangle_side}
\alias{triangle_side}
\title{Calculate Triangle Side Length with Pythagorean Theorem}
\usage{
triangle_side(side_a = NULL, side_b = NULL, side_c = NULL)
}
\arguments{
\item{side_a}{Numeric. The length of the first side of the triangle}

\item{side_b}{Numeric. The length of the second side of the triangle}

\item{side_c}{Numeric. The length of the hypotenuse (third) side of the triangle}
}
\value{
Numeric. The result of the pythagorean theorem calculating the missing side
}
\description{
This function calculates the length of the third side of a right triange,
given the lengths of the other two sides.
}
\examples{
Calculate the hypotenuse of a triangle
triangle_side(side_a = 3, side_b = 4)  #Returns 5
Calculate side b of a triangle
triangle_side(side_a = 3, side_c = 5)  #Returns 4
}
