% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimmed_mean.R
\name{trimmed_mean}
\alias{trimmed_mean}
\title{Calculates the Trimmed Mean}
\usage{
trimmed_mean(x = NULL, s = 0, l = 0)
}
\arguments{
\item{x}{A numeric vector. From this vector, the trimmed mean is calculated.}

\item{s}{Numeric. Number of smaller values to remove from the calculation.}

\item{l}{Numeric. Number of larger values to remove from the calculation.}
}
\value{
Numeric. The trimmed mean.
}
\description{
This function calculates the trimmed mean of a numeric vector
(e.g., mean that ignores the smallest and largest values).
}
\examples{
Calculate the trimmed mean
trimmed_mean(x = c(1, 7, 3, 2, 5, 0.5, 9, 10), s = 1, l = 2)  # Returns mean of c(1, 7, 3, 2, 5) = 3.6

}
